+++ trezorlib/ed25519raw.py
